Reset: move reset action in its own files
- extract a ui_reset.h
