+++ src/apdu_baking.c
@@ -32,39 +32,6 @@
 
 #include <string.h>
 
-#define G global.apdu.u.baking
-
-size_t handle_apdu_reset(__attribute__((unused)) uint8_t instruction, volatile uint32_t* flags) {
-    uint8_t* dataBuffer = G_io_apdu_buffer + OFFSET_CDATA;
-    uint32_t dataLength = G_io_apdu_buffer[OFFSET_LC];
-    if (dataLength != sizeof(level_t)) {
-        THROW(EXC_WRONG_LENGTH_FOR_INS);
-    }
-    level_t const lvl = READ_UNALIGNED_BIG_ENDIAN(level_t, dataBuffer);
-    if (!is_valid_level(lvl)) {
-        THROW(EXC_PARSE_ERROR);
-    }
-
-    G.reset_level = lvl;
-    ui_baking_reset(flags);
-    return 0;
-}
-
-bool reset_ok(void) {
-    UPDATE_NVRAM(ram, {
-        ram->hwm.main.highest_level = G.reset_level;
-        ram->hwm.main.highest_round = 0;
-        ram->hwm.main.had_endorsement = false;
-        ram->hwm.test.highest_level = G.reset_level;
-        ram->hwm.test.highest_round = 0;
-        ram->hwm.test.had_endorsement = false;
-    });
-
-    // Send back the response, do not restart the event loop
-    delayed_send(finalize_successful_send(0));
-    return true;
-}
-
 size_t send_word_big_endian(size_t tx, uint32_t word) {
     char word_bytes[sizeof(word)];
