+++ src/apdu_reset.h
@@ -1,8 +1,9 @@
-/* Tezos Ledger application - Baking BAGL UI handling
+/* Tezos Ledger application - Reset APDU instruction handling
 
    Copyright 2024 TriliTech <contact@trili.tech>
    Copyright 2024 Functori <contact@functori.com>
    Copyright 2023 Ledger
+   Copyright 2022 Nomadic Labs <contact@nomadic-labs.com>
    Copyright 2019 Obsidian Systems
 
    Licensed under the Apache License, Version 2.0 (the "License");
@@ -18,28 +19,11 @@
    limitations under the License.
 
 */
+#pragma once
 
-#ifdef HAVE_BAGL
+#include <stddef.h>
+#include <stdint.h>
+#include <stdbool.h>
 
-#include "apdu_baking.h"
-
-#include "apdu.h"
-#include "baking_auth.h"
-#include "globals.h"
-#include "os_cx.h"
-#include "protocol.h"
-#include "to_string.h"
-#include "ui.h"
-
-#include <string.h>
-
-#define G global.apdu.u.baking
-
-void ui_baking_reset(__attribute__((unused)) volatile uint32_t* flags) {
-    init_screen_stack();
-    push_ui_callback("Reset HWM", number_to_string_indirect32, &G.reset_level);
-
-    ux_confirm_screen(reset_ok, delay_reject);
-}
-
-#endif  // HAVE_BAGL
+bool reset_ok(void);
+size_t handle_apdu_reset(uint8_t instruction, volatile uint32_t* flags);
