+++ src/ui_reset.h
@@ -1,4 +1,4 @@
-/* Tezos Ledger application - Baking APDU instruction handling
+/* Tezos Ledger application - Reset UI handling
 
    Copyright 2024 TriliTech <contact@trili.tech>
    Copyright 2024 Functori <contact@functori.com>
@@ -19,15 +19,8 @@
    limitations under the License.
 
 */
-
 #pragma once
 
-#include <stddef.h>
-#include <stdint.h>
-#include <stdbool.h>
+#include "types.h"
 
-size_t handle_apdu_query_auth_key(uint8_t instruction, volatile uint32_t* flags);
-size_t handle_apdu_query_auth_key_with_curve(uint8_t instruction, volatile uint32_t* flags);
-size_t handle_apdu_main_hwm(uint8_t instruction, volatile uint32_t* flags);
-size_t handle_apdu_all_hwm(uint8_t instruction, volatile uint32_t* flags);
-size_t handle_apdu_deauthorize(uint8_t instruction, volatile uint32_t* flags);
+void ui_baking_reset(volatile uint32_t* flags);
